isolate algorithm options in separate files
