+++ .gitignore
@@ -14,3 +14,4 @@ imgui.ini
 CMakeSettings.json
 .jekyll-cache
 _site
+exiv2/
