+++ CMakeLists.txt
@@ -33,6 +33,7 @@ set(XPANO_SOURCES
   "xpano/algorithm/auto_crop.cc"
   "xpano/algorithm/bundle_adjuster.cc"
   "xpano/algorithm/image.cc"
+  "xpano/algorithm/options.cc"
   "xpano/log/logger.cc"
   "xpano/gui/backends/base.cc"
   "xpano/gui/backends/sdl.cc"
@@ -47,6 +48,7 @@ set(XPANO_SOURCES
   "xpano/gui/panels/warning_pane.cc"
   "xpano/gui/pano_gui.cc"
   "xpano/gui/shortcut.cc"
+  "xpano/pipeline/options.cc"
   "xpano/pipeline/stitcher_pipeline.cc"
   "xpano/utils/config.cc"
   "xpano/utils/disjoint_set.cc"
