+++ xpano/pipeline/stitcher_pipeline.h
@@ -13,41 +13,10 @@
 #include "xpano/algorithm/algorithm.h"
 #include "xpano/algorithm/image.h"
 #include "xpano/constants.h"
+#include "xpano/pipeline/options.h"
 #include "xpano/utils/rect.h"
 
 namespace xpano::pipeline {
-
-using InpaintingOptions = algorithm::InpaintingOptions;
-using StitchAlgorithmOptions = algorithm::StitchOptions;
-
-enum class ChromaSubsampling {
-  k444,
-  k422,
-  k420,
-};
-
-const auto kSubsamplingModes = std::array{
-    ChromaSubsampling::k444, ChromaSubsampling::k422, ChromaSubsampling::k420};
-
-const char *Label(ChromaSubsampling subsampling);
-
-struct CompressionOptions {
-  int jpeg_quality = kDefaultJpegQuality;
-  bool jpeg_progressive = false;
-  bool jpeg_optimize = false;
-  ChromaSubsampling jpeg_subsampling = ChromaSubsampling::k422;
-  int png_compression = kDefaultPngCompression;
-};
-
-struct MatchingOptions {
-  int neighborhood_search_size = kDefaultNeighborhoodSearchSize;
-  int match_threshold = kDefaultMatchThreshold;
-  float match_conf = kDefaultMatchConf;
-};
-
-struct LoadingOptions {
-  int preview_longer_side = kDefaultPreviewLongerSide;
-};
 
 struct StitchingOptions {
   int pano_id = 0;
